requirements.txt
python-dotenv~=1.0.0
requests~=2.31.0
bs4~=0.0.1
beautifulsoup4~=4.12.2
jsonlines~=3.1.0